Rename project to MyInflucDbClient
