+++ src/MyInfluxDbClient.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InfluxDbClient", "projects\InfluxDbClient\InfluxDbClient.csproj", "{8BDC0C80-D83B-42EB-915B-A1694559E16A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyInfluxDbClient", "projects\MyInfluxDbClient\MyInfluxDbClient.csproj", "{8BDC0C80-D83B-42EB-915B-A1694559E16A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{FBA435B4-A054-4AFB-88F8-EAD82C78915E}"
 	ProjectSection(SolutionItems) = preProject
