+++ src/projects/MyInfluxDbClient/MicrosecondsResolution.cs
@@ -1,4 +1,4 @@
-﻿namespace InfluxDbClient
+﻿namespace MyInfluxDbClient
 {
     public class MicrosecondsResolution : TimeStampResolution
     {
