+++ src/projects/MyInfluxDbClient/NanosecondsResolution.cs
@@ -1,4 +1,4 @@
-﻿namespace InfluxDbClient
+﻿namespace MyInfluxDbClient
 {
     public class NanosecondsResolution : TimeStampResolution
     {
