+++ src/projects/MyInfluxDbClient/Properties/AssemblyInfo.cs
