+++ src/projects/MyInfluxDbClient/TimeStamp.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace InfluxDbClient
+namespace MyInfluxDbClient
 {
     public struct TimeStamp
     {
