+++ src/projects/MyInfluxDbClient/TimeStampResolution.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace InfluxDbClient
+namespace MyInfluxDbClient
 {
     public abstract class TimeStampResolution
     {
