+++ src/projects/MyInfluxDbClient/packages.config
