+++ src/samples/MrLab/MrLab.csproj
@@ -48,9 +48,9 @@
     <None Include="App.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\projects\InfluxDbClient\InfluxDbClient.csproj">
+    <ProjectReference Include="..\..\projects\MyInfluxDbClient\MyInfluxDbClient.csproj">
       <Project>{8bdc0c80-d83b-42eb-915b-a1694559e16a}</Project>
-      <Name>InfluxDbClient</Name>
+      <Name>MyInfluxDbClient</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
